windows: prevent inclusion of winsock.h in another way
The other solution DID prevent the inclusion of winsock but still lead to a macro-redefinition-warning when including winsock2. And because we are treating warnings as errors, it broke the build.
Now, we can either disable the warning or not include windows.h. Oooor make windows.h include less overhead. Which I just did.

It might potentially pose an issue when using the biotracker lib with other libs that rely on windows.h including stuff like cryptography. Then again: those are probably rare, they probably have a clean interface and we are linking them as a DLL.
So yeah, here you go until someone complains.

More information:
https://blogs.msdn.microsoft.com/oldnewthing/20091130-00/?p=15863